reordering components folder
