--- src/app/components/CreateEventComponent.tsx
@@ -1,113 +1,0 @@
-'use client'
-
-import { useState } from 'react'
-import ComboboxCountries from './ComboboxCountries/ComboboxCountries'
-import styles from './CreateEventComponent.module.css'
-import ListOfCountries from './ListOfCoutries/ListOfCountries'
-
-const CreateEventComponent = () => {
-  const [isOpenSelectTimeZone, setIsOpenSelectTimeZone] = useState(false)
-
-  const handleClose = () => {
-    setIsOpenSelectTimeZone(false)
-  }
-
-  return (
-    <>
-      <div className={styles['container-form']}>
-        <form action="" className={styles.form}>
-          <div>
-            <input
-              type="text"
-              name=""
-              id=""
-              className={styles['full-width']}
-              placeholder="Event name"
-            />
-          </div>
-
-          <div className={styles['container-half-width']}>
-            <div>
-              <input
-                type="time"
-                name=""
-                id=""
-                className={styles['full-width']}
-              />
-            </div>
-            <div>
-              <input
-                type="date"
-                name=""
-                id=""
-                className={styles['full-width']}
-              />
-            </div>
-          </div>
-
-          <div className={styles['container-half-width']}>
-            <div className={styles.halfWidth}>
-              {/* Acá se abre el modal de los timeZone */}
-              <button
-                type="button"
-                onClick={() => setIsOpenSelectTimeZone(true)}
-              >
-                Time zone
-              </button>
-              {isOpenSelectTimeZone && (
-                <ListOfCountries onClose={handleClose} />
-              )}
-            </div>
-            <div className={styles.halfWidth}>
-              <select name="" id="" className={styles['full-width']}>
-                <option value="lg-1">First language</option>
-                <option value="lg-2">Second language</option>
-              </select>
-            </div>
-          </div>
-
-          <div>
-            <input
-              type="url"
-              name=""
-              id=""
-              placeholder="Hyperlink"
-              className={styles['full-width']}
-            />
-          </div>
-          <div className={styles.prueba}>
-            {/* // TODO: El div que envuelve los dos textarea, tiene como 2-3 más de altura | Help */}
-            <textarea
-              name=""
-              id=""
-              placeholder="Description"
-              className={styles['full-width']}
-            />
-          </div>
-          <div>
-            {/* //! TODO: how to do an update image? */}
-            <input
-              type="text"
-              name=""
-              id=""
-              placeholder="how to do an update image?"
-              className={styles['full-width']}
-            />
-          </div>
-          <ComboboxCountries format={24} />
-          <div>
-            {/* // TODO: El div que envuelve los dos textarea, tiene como 2-3 más de altura | Help */}
-            <textarea
-              name=""
-              id=""
-              placeholder="🚩 🚩 🚩 🚩 🚩 🚩 🚩 🚩"
-              className={styles['full-width']}
-            />
-          </div>
-        </form>
-      </div>
-    </>
-  )
-}
-
-export default CreateEventComponent
