+++ src/app/list-countries/page.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { useState } from 'react'
-import ListOfCountries from '../components/ListOfCoutries/ListOfCountries'
+import CountryList from '@/app/components/molecules/country-list/CountryList'
 
 const ListCountriesPage = () => {
   const [isOpenSelectTimeZone, setIsOpenSelectTimeZone] = useState(false)
@@ -10,13 +10,13 @@ const ListCountriesPage = () => {
     setIsOpenSelectTimeZone(false)
   }
 
-// ! Esto es una página para hacer prubas del modal de TimeZone,
-// ! terminará siendo eliminada
+  // ! Esto es una página para hacer prubas del modal de TimeZone,
+  // ! terminará siendo eliminada
 
   return (
     <>
       <button onClick={() => setIsOpenSelectTimeZone(true)}>Time zone</button>
-      {isOpenSelectTimeZone && <ListOfCountries onClose={handleClose} />}
+      {isOpenSelectTimeZone && <CountryList onClose={handleClose} />}
     </>
   )
 }
